getResizeMatrix + cleanup
Gave GraphicsObjectManager the matrix for scaling and translating points
so that they stay at the same "pixel position" when the window is
resized. (Used to be in Arrow.h)

Also setting the resize matrix uniform on window resize instead of every
frame (in Arrow). This does make the HUD objects seem like they're
"lagging behind" when continually resizing the window, but that's not
the intended usage. It improves performance slightly as well.