Static EventHandler
Removed the weird semi-static state of EventHandler in favor of a pure
static state. This will make it easier to access its hook functions,
like in destructors, where arguments cannot be passed.

As an end note, I do not think the now global nature of EventHandler is
going to be a problem in the future.